Added some rudimentary path blocks for the tranquility altar - textures pending.

+++ src/main/java/WayofTime/bloodmagic/api/util/helper/IncenseHelper.java
@@ -6,20 +6,20 @@ import net.minecraft.nbt.NBTTagCompound;
 
 public class IncenseHelper
 {
-    public static float getCurrentIncense(EntityPlayer player)
+    public static double getCurrentIncense(EntityPlayer player)
     {
         NBTTagCompound data = player.getEntityData();
         if (data.hasKey(Constants.NBT.CURRENT_INCENSE))
         {
-            return data.getFloat(Constants.NBT.CURRENT_INCENSE);
+            return data.getDouble(Constants.NBT.CURRENT_INCENSE);
         }
 
         return 0;
     }
 
-    public static void setCurrentIncense(EntityPlayer player, float amount)
+    public static void setCurrentIncense(EntityPlayer player, double amount)
     {
         NBTTagCompound data = player.getEntityData();
-        data.setFloat(Constants.NBT.CURRENT_INCENSE, amount);
+        data.setDouble(Constants.NBT.CURRENT_INCENSE, amount);
     }
 }
